+++ packages/fictoan-docs/src/app/page.client.tsx
@@ -83,11 +83,11 @@ const HomePage = () => {
                             language="bash"
                             withSyntaxHighlighting
                             showCopyButton
-                            source={`pnpm add fictoan-react@2.0.0-beta.5
+                            source={`pnpm add fictoan-react@2.0.0-beta.6
 # or
-yarn add fictoan-react@2.0.0-beta.5
+yarn add fictoan-react@2.0.0-beta.6
 # or
-npm install fictoan-react@2.0.0-beta.5`}
+npm install fictoan-react@2.0.0-beta.6`}
                         />
                     </Portion>